compile selected kernels with gcc and clang
This commit provides Docker images, scripts and documentation to set
up the system to compile the Linux kernel with gcc and clang.

It serves as starting point to do some analysis of the output of
compiler warnings, static analysis tools in clang, and as a scheme
to set up more sophisticated tools for analysis of the Linux kernel.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>