compile with clang-5.0, clang-6.0 or Debian's unstable clang-7
As compiling the kernel with clang is a recent development, we may
observe different failures between the clang compiler versions.

Some kernel configurations may fail with the stable version 5.0
due to clang compiler shortcomings, that have been resolved in the
later version 6.0 or in the svn tree of the clang development.

Hence, this commit provides dedicated Docker containers with
clang-6.0 and clang-7 from Debian's unstable distribution, and
extends the scripts select among those clang compiler versions.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>