scripts: compile kernel in docker as user instead of root
By default, the docker containers executes with root, but then,
the intermediate files placed in the volume cannot be removed
afterwards by the user. So, we need to execute the build commands
as the user calling the script.

Furthermore, the kernel build uses `whoami`; so, we need to create
the user in the docker container with the correct id and user name.

Signed-off-by: Ozan Alpay <ozanalpay@yandex.com.tr>
[lukas.bulwahn@gmail.com: rewrote the commit message]
Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>